+++ First game/src/Player.java
@@ -52,12 +52,12 @@ public class Player extends GameObject{
 	public void render(Graphics g) {
 		if(id == ID.Player) {
 		g.setColor(Color.magenta);//use color class to set color
-		g.fillRect(x, y, 32, 32);//this way always same size but place determined by fields, rn it's a square of 32 by 32
+		g.fillRect((int)x,(int)y, 32, 32);//this way always same size but place determined by fields, rn it's a square of 32 by 32
 		}
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 32, 32);//sets bounds to same as the player
+		return new Rectangle((int)x,(int)y, 32, 32);//sets bounds to same as the player
 	}
